requirements-frozen-linux.txt
# Common
ansible==8.3.0; sys_platform != 'win32'
batcave==42.2.0
PyYAML==6.0.1
dicttoxml==1.7.16
dotmap==1.3.30
# Test
ansible-lint==6.19.0; sys_platform != 'win32'
autopep8==2.0.4
doc8==1.1.1
flake8==6.1.0
flake8-annotations==3.0.1
Flake8-pyproject==1.2.3
junitparser==3.1.0
mypy==1.5.1
pylint==2.17.5
restview==3.0.1
rope==1.9.0
types-PyYAML==6.0.12.11
types-requests==2.31.0.2
types-setuptools==68.2.0.0
types-xmltodict==0.13.0.3
unittest-xml-reporting==3.2.0
# Deploy
pywinrm==0.4.3
# Release
jira==3.5.2
twine==4.0.2
## The following requirements were added by pip freeze:
ansible-compat==4.1.10; sys_platform != 'win32'
ansible-core==2.15.3; sys_platform != 'win32'
astroid==2.15.6
attrs==23.1.0
black==23.7.0
bleach==6.0.0
bracex==2.4
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
dill==0.3.7
docker==6.1.3
docutils==0.20.1
filelock==3.12.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.35
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==24.2.0
kubernetes==26.1.0
lazy-object-proxy==1.9.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
pbr==5.11.1
pkginfo==1.9.6
platformdirs==3.10.0
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyspnego==0.9.2
python-dateutil==2.8.2
pytoolconfig==1.2.5
readme-renderer==36.0
referencing==0.30.2
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
resolvelib==1.0.1
restructuredtext-lint==1.4.0
rfc3986==2.0.0
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.3
six==1.16.0
smmap==5.0.0
stevedore==5.1.0
subprocess-tee==0.4.1
tomlkit==0.12.1
types-urllib3==1.26.25.14
typing_extensions==4.7.1
urllib3==1.26.16
wcmatch==8.5
webencodings==0.5.1
websocket-client==1.6.2
wrapt==1.15.0
xmltodict==0.13.0
yamllint==1.32.0
zipp==3.16.2
